requirements.txt
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
python-dotenv>=1.0.0
mss>=9.0.0
numpy>=2.0.0
Pillow>=11.0.0
pyaudio>=0.2.14
pyobjc-core>=10.0
pyobjc-framework-Vision>=10.0
pyobjc-framework-Quartz>=10.0
thefuzz>=0.22.0
python-Levenshtein>=0.26.0
python-multipart>=0.0.18
pynput>=1.7.7
